+++ sdk/resourcemanager/devcenter/armdevcenter/models.go
@@ -114,14 +114,14 @@ type AttachedNetworksClientGetByProjectOptions struct {
 	// placeholder for future optional parameters
 }
 
-// AttachedNetworksClientListByDevCenterOptions contains the optional parameters for the AttachedNetworksClient.ListByDevCenter
+// AttachedNetworksClientListByDevCenterOptions contains the optional parameters for the AttachedNetworksClient.NewListByDevCenterPager
 // method.
 type AttachedNetworksClientListByDevCenterOptions struct {
 	// The maximum number of resources to return from the operation. Example: '$top=10'.
 	Top *int32
 }
 
-// AttachedNetworksClientListByProjectOptions contains the optional parameters for the AttachedNetworksClient.ListByProject
+// AttachedNetworksClientListByProjectOptions contains the optional parameters for the AttachedNetworksClient.NewListByProjectPager
 // method.
 type AttachedNetworksClientListByProjectOptions struct {
 	// The maximum number of resources to return from the operation. Example: '$top=10'.
@@ -230,7 +230,7 @@ type CatalogsClientGetOptions struct {
 	// placeholder for future optional parameters
 }
 
-// CatalogsClientListByDevCenterOptions contains the optional parameters for the CatalogsClient.ListByDevCenter method.
+// CatalogsClientListByDevCenterOptions contains the optional parameters for the CatalogsClient.NewListByDevCenterPager method.
 type CatalogsClientListByDevCenterOptions struct {
 	// The maximum number of resources to return from the operation. Example: '$top=10'.
 	Top *int32
@@ -385,14 +385,14 @@ type DevBoxDefinitionsClientGetOptions struct {
 	// placeholder for future optional parameters
 }
 
-// DevBoxDefinitionsClientListByDevCenterOptions contains the optional parameters for the DevBoxDefinitionsClient.ListByDevCenter
+// DevBoxDefinitionsClientListByDevCenterOptions contains the optional parameters for the DevBoxDefinitionsClient.NewListByDevCenterPager
 // method.
 type DevBoxDefinitionsClientListByDevCenterOptions struct {
 	// The maximum number of resources to return from the operation. Example: '$top=10'.
 	Top *int32
 }
 
-// DevBoxDefinitionsClientListByProjectOptions contains the optional parameters for the DevBoxDefinitionsClient.ListByProject
+// DevBoxDefinitionsClientListByProjectOptions contains the optional parameters for the DevBoxDefinitionsClient.NewListByProjectPager
 // method.
 type DevBoxDefinitionsClientListByProjectOptions struct {
 	// The maximum number of resources to return from the operation. Example: '$top=10'.
@@ -450,18 +450,36 @@ type DevCentersClientGetOptions struct {
 	// placeholder for future optional parameters
 }
 
-// DevCentersClientListByResourceGroupOptions contains the optional parameters for the DevCentersClient.ListByResourceGroup
+// DevCentersClientListByResourceGroupOptions contains the optional parameters for the DevCentersClient.NewListByResourceGroupPager
 // method.
 type DevCentersClientListByResourceGroupOptions struct {
 	// The maximum number of resources to return from the operation. Example: '$top=10'.
 	Top *int32
 }
 
-// DevCentersClientListBySubscriptionOptions contains the optional parameters for the DevCentersClient.ListBySubscription
+// DevCentersClientListBySubscriptionOptions contains the optional parameters for the DevCentersClient.NewListBySubscriptionPager
 // method.
 type DevCentersClientListBySubscriptionOptions struct {
 	// The maximum number of resources to return from the operation. Example: '$top=10'.
 	Top *int32
+}
+
+// EndpointDependency - A domain name and connection details used to access a dependency.
+type EndpointDependency struct {
+	// READ-ONLY; Human-readable supplemental information about the dependency and when it is applicable.
+	Description *string `json:"description,omitempty" azure:"ro"`
+
+	// READ-ONLY; The domain name of the dependency. Domain names may be fully qualified or may contain a * wildcard.
+	DomainName *string `json:"domainName,omitempty" azure:"ro"`
+
+	// READ-ONLY; The list of connection details for this endpoint.
+	EndpointDetails []*EndpointDetail `json:"endpointDetails,omitempty" azure:"ro"`
+}
+
+// EndpointDetail - Details about the connection between the Batch service and the endpoint.
+type EndpointDetail struct {
+	// READ-ONLY; The port an endpoint is connected to.
+	Port *int32 `json:"port,omitempty" azure:"ro"`
 }
 
 // EnvironmentRole - A role that can be assigned to a user.
@@ -532,7 +550,7 @@ type EnvironmentTypesClientGetOptions struct {
 	// placeholder for future optional parameters
 }
 
-// EnvironmentTypesClientListByDevCenterOptions contains the optional parameters for the EnvironmentTypesClient.ListByDevCenter
+// EnvironmentTypesClientListByDevCenterOptions contains the optional parameters for the EnvironmentTypesClient.NewListByDevCenterPager
 // method.
 type EnvironmentTypesClientListByDevCenterOptions struct {
 	// The maximum number of resources to return from the operation. Example: '$top=10'.
@@ -547,7 +565,7 @@ type EnvironmentTypesClientUpdateOptions struct {
 // ErrorAdditionalInfo - The resource management error additional info.
 type ErrorAdditionalInfo struct {
 	// READ-ONLY; The additional info.
-	Info interface{} `json:"info,omitempty" azure:"ro"`
+	Info any `json:"info,omitempty" azure:"ro"`
 
 	// READ-ONLY; The additional info type.
 	Type *string `json:"type,omitempty" azure:"ro"`
@@ -589,7 +607,8 @@ type GalleriesClientGetOptions struct {
 	// placeholder for future optional parameters
 }
 
-// GalleriesClientListByDevCenterOptions contains the optional parameters for the GalleriesClient.ListByDevCenter method.
+// GalleriesClientListByDevCenterOptions contains the optional parameters for the GalleriesClient.NewListByDevCenterPager
+// method.
 type GalleriesClientListByDevCenterOptions struct {
 	// The maximum number of resources to return from the operation. Example: '$top=10'.
 	Top *int32
@@ -709,6 +728,15 @@ type HealthCheckStatusDetailsProperties struct {
 	StartDateTime *time.Time `json:"startDateTime,omitempty" azure:"ro"`
 }
 
+// HealthStatusDetail - Pool health status detail.
+type HealthStatusDetail struct {
+	// READ-ONLY; An identifier for the issue.
+	Code *string `json:"code,omitempty" azure:"ro"`
+
+	// READ-ONLY; A message describing the issue, intended to be suitable for display in a user interface
+	Message *string `json:"message,omitempty" azure:"ro"`
+}
+
 // Image - Represents an image.
 type Image struct {
 	// Image properties.
@@ -741,6 +769,10 @@ type ImageProperties struct {
 	// READ-ONLY; The description of the image.
 	Description *string `json:"description,omitempty" azure:"ro"`
 
+	// READ-ONLY; Indicates whether this image has hibernate enabled. Not all images are capable of supporting hibernation. To
+	// find out more see https://aka.ms/devbox/hibernate
+	HibernateSupport *HibernateSupport `json:"hibernateSupport,omitempty" azure:"ro"`
+
 	// READ-ONLY; The name of the image offer.
 	Offer *string `json:"offer,omitempty" azure:"ro"`
 
@@ -761,15 +793,6 @@ type ImageProperties struct {
 type ImageReference struct {
 	// Image ID, or Image version ID. When Image ID is provided, its latest version will be used.
 	ID *string `json:"id,omitempty"`
-
-	// The image offer.
-	Offer *string `json:"offer,omitempty"`
-
-	// The image publisher.
-	Publisher *string `json:"publisher,omitempty"`
-
-	// The image sku.
-	SKU *string `json:"sku,omitempty"`
 
 	// READ-ONLY; The actual version of the image after use. When id references a gallery image latest version, this will indicate
 	// the actual version in use.
@@ -835,7 +858,8 @@ type ImageVersionsClientGetOptions struct {
 	// placeholder for future optional parameters
 }
 
-// ImageVersionsClientListByImageOptions contains the optional parameters for the ImageVersionsClient.ListByImage method.
+// ImageVersionsClientListByImageOptions contains the optional parameters for the ImageVersionsClient.NewListByImagePager
+// method.
 type ImageVersionsClientListByImageOptions struct {
 	// placeholder for future optional parameters
 }
@@ -845,13 +869,13 @@ type ImagesClientGetOptions struct {
 	// placeholder for future optional parameters
 }
 
-// ImagesClientListByDevCenterOptions contains the optional parameters for the ImagesClient.ListByDevCenter method.
+// ImagesClientListByDevCenterOptions contains the optional parameters for the ImagesClient.NewListByDevCenterPager method.
 type ImagesClientListByDevCenterOptions struct {
 	// The maximum number of resources to return from the operation. Example: '$top=10'.
 	Top *int32
 }
 
-// ImagesClientListByGalleryOptions contains the optional parameters for the ImagesClient.ListByGallery method.
+// ImagesClientListByGalleryOptions contains the optional parameters for the ImagesClient.NewListByGalleryPager method.
 type ImagesClientListByGalleryOptions struct {
 	// The maximum number of resources to return from the operation. Example: '$top=10'.
 	Top *int32
@@ -864,15 +888,6 @@ type ListResult struct {
 
 	// READ-ONLY; Current page of results.
 	Value []*DevCenter `json:"value,omitempty" azure:"ro"`
-}
-
-// ListUsagesResult - List of Core Usages.
-type ListUsagesResult struct {
-	// READ-ONLY; The link to get the next page of Usage result.
-	NextLink *string `json:"nextLink,omitempty" azure:"ro"`
-
-	// READ-ONLY; The array page of Usages.
-	Value []*Usage `json:"value,omitempty" azure:"ro"`
 }
 
 // ManagedServiceIdentity - Managed service identity (system assigned and/or user assigned identities)
@@ -1000,23 +1015,30 @@ type NetworkConnectionsClientGetOptions struct {
 	// placeholder for future optional parameters
 }
 
-// NetworkConnectionsClientListByResourceGroupOptions contains the optional parameters for the NetworkConnectionsClient.ListByResourceGroup
+// NetworkConnectionsClientListByResourceGroupOptions contains the optional parameters for the NetworkConnectionsClient.NewListByResourceGroupPager
 // method.
 type NetworkConnectionsClientListByResourceGroupOptions struct {
 	// The maximum number of resources to return from the operation. Example: '$top=10'.
 	Top *int32
 }
 
-// NetworkConnectionsClientListBySubscriptionOptions contains the optional parameters for the NetworkConnectionsClient.ListBySubscription
+// NetworkConnectionsClientListBySubscriptionOptions contains the optional parameters for the NetworkConnectionsClient.NewListBySubscriptionPager
 // method.
 type NetworkConnectionsClientListBySubscriptionOptions struct {
 	// The maximum number of resources to return from the operation. Example: '$top=10'.
 	Top *int32
 }
 
-// NetworkConnectionsClientListHealthDetailsOptions contains the optional parameters for the NetworkConnectionsClient.ListHealthDetails
+// NetworkConnectionsClientListHealthDetailsOptions contains the optional parameters for the NetworkConnectionsClient.NewListHealthDetailsPager
 // method.
 type NetworkConnectionsClientListHealthDetailsOptions struct {
+	// The maximum number of resources to return from the operation. Example: '$top=10'.
+	Top *int32
+}
+
+// NetworkConnectionsClientListOutboundNetworkDependenciesEndpointsOptions contains the optional parameters for the NetworkConnectionsClient.NewListOutboundNetworkDependenciesEndpointsPager
+// method.
+type NetworkConnectionsClientListOutboundNetworkDependenciesEndpointsOptions struct {
 	// The maximum number of resources to return from the operation. Example: '$top=10'.
 	Top *int32
 }
@@ -1129,7 +1151,7 @@ type OperationStatus struct {
 	StartTime *time.Time `json:"startTime,omitempty"`
 
 	// READ-ONLY; Custom operation properties, populated only for a successful operation.
-	Properties interface{} `json:"properties,omitempty" azure:"ro"`
+	Properties any `json:"properties,omitempty" azure:"ro"`
 
 	// READ-ONLY; The id of the resource.
 	ResourceID *string `json:"resourceId,omitempty" azure:"ro"`
@@ -1167,9 +1189,28 @@ type OperationStatusesClientGetOptions struct {
 	// placeholder for future optional parameters
 }
 
-// OperationsClientListOptions contains the optional parameters for the OperationsClient.List method.
+// OperationsClientListOptions contains the optional parameters for the OperationsClient.NewListPager method.
 type OperationsClientListOptions struct {
 	// placeholder for future optional parameters
+}
+
+// OutboundEnvironmentEndpoint - A collection of related endpoints from the same service for which the agent requires outbound
+// access.
+type OutboundEnvironmentEndpoint struct {
+	// READ-ONLY; The type of service that the agent connects to.
+	Category *string `json:"category,omitempty" azure:"ro"`
+
+	// READ-ONLY; The endpoints for this service for which the agent requires outbound access.
+	Endpoints []*EndpointDependency `json:"endpoints,omitempty" azure:"ro"`
+}
+
+// OutboundEnvironmentEndpointCollection - Values returned by the List operation.
+type OutboundEnvironmentEndpointCollection struct {
+	// The continuation token.
+	NextLink *string `json:"nextLink,omitempty"`
+
+	// READ-ONLY; The collection of outbound network dependency endpoints returned by the listing operation.
+	Value []*OutboundEnvironmentEndpoint `json:"value,omitempty" azure:"ro"`
 }
 
 // Pool - A pool of Virtual Machines.
@@ -1219,6 +1260,16 @@ type PoolProperties struct {
 	// Name of a Network Connection in parent Project of this Pool
 	NetworkConnectionName *string `json:"networkConnectionName,omitempty"`
 
+	// Stop on disconnect configuration settings for Dev Boxes created in this pool.
+	StopOnDisconnect *StopOnDisconnectConfiguration `json:"stopOnDisconnect,omitempty"`
+
+	// READ-ONLY; Overall health status of the Pool. Indicates whether or not the Pool is available to create Dev Boxes.
+	HealthStatus *HealthStatus `json:"healthStatus,omitempty" azure:"ro"`
+
+	// READ-ONLY; Details on the Pool health status to help diagnose issues. This is only populated when the pool status indicates
+	// the pool is in a non-healthy state
+	HealthStatusDetails []*HealthStatusDetail `json:"healthStatusDetails,omitempty" azure:"ro"`
+
 	// READ-ONLY; The provisioning state of the resource.
 	ProvisioningState *ProvisioningState `json:"provisioningState,omitempty" azure:"ro"`
 }
@@ -1248,6 +1299,9 @@ type PoolUpdateProperties struct {
 
 	// Name of a Network Connection in parent Project of this Pool
 	NetworkConnectionName *string `json:"networkConnectionName,omitempty"`
+
+	// Stop on disconnect configuration settings for Dev Boxes created in this pool.
+	StopOnDisconnect *StopOnDisconnectConfiguration `json:"stopOnDisconnect,omitempty"`
 }
 
 // PoolsClientBeginCreateOrUpdateOptions contains the optional parameters for the PoolsClient.BeginCreateOrUpdate method.
@@ -1258,6 +1312,12 @@ type PoolsClientBeginCreateOrUpdateOptions struct {
 
 // PoolsClientBeginDeleteOptions contains the optional parameters for the PoolsClient.BeginDelete method.
 type PoolsClientBeginDeleteOptions struct {
+	// Resumes the LRO from the provided token.
+	ResumeToken string
+}
+
+// PoolsClientBeginRunHealthChecksOptions contains the optional parameters for the PoolsClient.BeginRunHealthChecks method.
+type PoolsClientBeginRunHealthChecksOptions struct {
 	// Resumes the LRO from the provided token.
 	ResumeToken string
 }
@@ -1273,7 +1333,7 @@ type PoolsClientGetOptions struct {
 	// placeholder for future optional parameters
 }
 
-// PoolsClientListByProjectOptions contains the optional parameters for the PoolsClient.ListByProject method.
+// PoolsClientListByProjectOptions contains the optional parameters for the PoolsClient.NewListByProjectPager method.
 type PoolsClientListByProjectOptions struct {
 	// The maximum number of resources to return from the operation. Example: '$top=10'.
 	Top *int32
@@ -1309,7 +1369,7 @@ type ProjectAllowedEnvironmentTypesClientGetOptions struct {
 	// placeholder for future optional parameters
 }
 
-// ProjectAllowedEnvironmentTypesClientListOptions contains the optional parameters for the ProjectAllowedEnvironmentTypesClient.List
+// ProjectAllowedEnvironmentTypesClientListOptions contains the optional parameters for the ProjectAllowedEnvironmentTypesClient.NewListPager
 // method.
 type ProjectAllowedEnvironmentTypesClientListOptions struct {
 	// The maximum number of resources to return from the operation. Example: '$top=10'.
@@ -1362,7 +1422,7 @@ type ProjectEnvironmentTypeProperties struct {
 	DeploymentTargetID *string `json:"deploymentTargetId,omitempty"`
 
 	// Defines whether this Environment Type can be used in this Project.
-	Status *EnableStatus `json:"status,omitempty"`
+	Status *EnvironmentTypeEnableStatus `json:"status,omitempty"`
 
 	// Role Assignments created on environment backing resources. This is a mapping from a user object ID to an object of role
 	// definition IDs.
@@ -1396,7 +1456,7 @@ type ProjectEnvironmentTypeUpdateProperties struct {
 	DeploymentTargetID *string `json:"deploymentTargetId,omitempty"`
 
 	// Defines whether this Environment Type can be used in this Project.
-	Status *EnableStatus `json:"status,omitempty"`
+	Status *EnvironmentTypeEnableStatus `json:"status,omitempty"`
 
 	// Role Assignments created on environment backing resources. This is a mapping from a user object ID to an object of role
 	// definition IDs.
@@ -1427,7 +1487,8 @@ type ProjectEnvironmentTypesClientGetOptions struct {
 	// placeholder for future optional parameters
 }
 
-// ProjectEnvironmentTypesClientListOptions contains the optional parameters for the ProjectEnvironmentTypesClient.List method.
+// ProjectEnvironmentTypesClientListOptions contains the optional parameters for the ProjectEnvironmentTypesClient.NewListPager
+// method.
 type ProjectEnvironmentTypesClientListOptions struct {
 	// The maximum number of resources to return from the operation. Example: '$top=10'.
 	Top *int32
@@ -1456,6 +1517,10 @@ type ProjectProperties struct {
 	// Resource Id of an associated DevCenter
 	DevCenterID *string `json:"devCenterId,omitempty"`
 
+	// When specified, limits the maximum number of Dev Boxes a single user can create across all pools in the project. This will
+	// have no effect on existing Dev Boxes when reduced.
+	MaxDevBoxesPerUser *int32 `json:"maxDevBoxesPerUser,omitempty"`
+
 	// READ-ONLY; The URI of the resource.
 	DevCenterURI *string `json:"devCenterUri,omitempty" azure:"ro"`
 
@@ -1482,6 +1547,10 @@ type ProjectUpdateProperties struct {
 
 	// Resource Id of an associated DevCenter
 	DevCenterID *string `json:"devCenterId,omitempty"`
+
+	// When specified, limits the maximum number of Dev Boxes a single user can create across all pools in the project. This will
+	// have no effect on existing Dev Boxes when reduced.
+	MaxDevBoxesPerUser *int32 `json:"maxDevBoxesPerUser,omitempty"`
 }
 
 // ProjectsClientBeginCreateOrUpdateOptions contains the optional parameters for the ProjectsClient.BeginCreateOrUpdate method.
@@ -1507,13 +1576,15 @@ type ProjectsClientGetOptions struct {
 	// placeholder for future optional parameters
 }
 
-// ProjectsClientListByResourceGroupOptions contains the optional parameters for the ProjectsClient.ListByResourceGroup method.
+// ProjectsClientListByResourceGroupOptions contains the optional parameters for the ProjectsClient.NewListByResourceGroupPager
+// method.
 type ProjectsClientListByResourceGroupOptions struct {
 	// The maximum number of resources to return from the operation. Example: '$top=10'.
 	Top *int32
 }
 
-// ProjectsClientListBySubscriptionOptions contains the optional parameters for the ProjectsClient.ListBySubscription method.
+// ProjectsClientListBySubscriptionOptions contains the optional parameters for the ProjectsClient.NewListBySubscriptionPager
+// method.
 type ProjectsClientListBySubscriptionOptions struct {
 	// The maximum number of resources to return from the operation. Example: '$top=10'.
 	Top *int32
@@ -1604,7 +1675,7 @@ type SKUListResult struct {
 	Value []*SKUInfo `json:"value,omitempty" azure:"ro"`
 }
 
-// SKUsClientListBySubscriptionOptions contains the optional parameters for the SKUsClient.ListBySubscription method.
+// SKUsClientListBySubscriptionOptions contains the optional parameters for the SKUsClient.NewListBySubscriptionPager method.
 type SKUsClientListBySubscriptionOptions struct {
 	// The maximum number of resources to return from the operation. Example: '$top=10'.
 	Top *int32
@@ -1643,7 +1714,7 @@ type ScheduleProperties struct {
 	Frequency *ScheduledFrequency `json:"frequency,omitempty"`
 
 	// Indicates whether or not this scheduled task is enabled.
-	State *EnableStatus `json:"state,omitempty"`
+	State *ScheduleEnableStatus `json:"state,omitempty"`
 
 	// The target time to trigger the action. The format is HH:MM.
 	Time *string `json:"time,omitempty"`
@@ -1677,7 +1748,7 @@ type ScheduleUpdateProperties struct {
 	Frequency *ScheduledFrequency `json:"frequency,omitempty"`
 
 	// Indicates whether or not this scheduled task is enabled.
-	State *EnableStatus `json:"state,omitempty"`
+	State *ScheduleEnableStatus `json:"state,omitempty"`
 
 	// The target time to trigger the action. The format is HH:MM.
 	Time *string `json:"time,omitempty"`
@@ -1720,10 +1791,19 @@ type SchedulesClientGetOptions struct {
 	Top *int32
 }
 
-// SchedulesClientListByPoolOptions contains the optional parameters for the SchedulesClient.ListByPool method.
+// SchedulesClientListByPoolOptions contains the optional parameters for the SchedulesClient.NewListByPoolPager method.
 type SchedulesClientListByPoolOptions struct {
 	// The maximum number of resources to return from the operation. Example: '$top=10'.
 	Top *int32
+}
+
+// StopOnDisconnectConfiguration - Stop on disconnect configuration settings for Dev Boxes created in this pool.
+type StopOnDisconnectConfiguration struct {
+	// The specified time in minutes to wait before stopping a Dev Box once disconnect is detected.
+	GracePeriodMinutes *int32 `json:"gracePeriodMinutes,omitempty"`
+
+	// Whether the feature to stop the Dev Box on disconnect once the grace period has lapsed is enabled.
+	Status *StopOnDisconnectEnableStatus `json:"status,omitempty"`
 }
 
 // SystemData - Metadata pertaining to creation and last modification of the resource.
@@ -1757,35 +1837,6 @@ type Update struct {
 
 	// Resource tags.
 	Tags map[string]*string `json:"tags,omitempty"`
-}
-
-// Usage - The core usage details.
-type Usage struct {
-	// The current usage.
-	CurrentValue *int64 `json:"currentValue,omitempty"`
-
-	// The limit integer.
-	Limit *int64 `json:"limit,omitempty"`
-
-	// The name.
-	Name *UsageName `json:"name,omitempty"`
-
-	// The unit details.
-	Unit *UsageUnit `json:"unit,omitempty"`
-}
-
-// UsageName - The Usage Names.
-type UsageName struct {
-	// The localized name of the resource.
-	LocalizedValue *string `json:"localizedValue,omitempty"`
-
-	// The name of the resource.
-	Value *string `json:"value,omitempty"`
-}
-
-// UsagesClientListByLocationOptions contains the optional parameters for the UsagesClient.ListByLocation method.
-type UsagesClientListByLocationOptions struct {
-	// placeholder for future optional parameters
 }
 
 // UserAssignedIdentity - User assigned identity properties